feat(orientdb-core): make outV(), inV() to support class filter
